+++ classic-menu-in-navigation-block.php
@@ -3,7 +3,7 @@
  * Plugin Name:  Classic Menu in Navigation Block
  * Plugin URI:   https://www.webmandesign.eu/portfolio/classic-menu-in-navigation-block-wordpress-plugin/
  * Description:  Extending WordPress Navigation block with functionality to display classic menus.
- * Version:      1.0.2
+ * Version:      1.0.3
  * Author:       WebMan Design, Oliver Juhas
  * Author URI:   https://www.webmandesign.eu/
  * License:      GNU General Public License v3
@@ -29,13 +29,10 @@
 defined( 'ABSPATH' ) || exit;
 
 // Constants.
-define( 'CMINB_VERSION', '1.0.2' );
+define( 'CMINB_VERSION', '1.0.3' );
 define( 'CMINB_FILE', __FILE__ );
 define( 'CMINB_PATH', plugin_dir_path( CMINB_FILE ) ); // Trailing slashed.
 define( 'CMINB_URL', plugin_dir_url( CMINB_FILE ) ); // Trailing slashed.
-if ( ! defined( 'CMINB_USE_CACHE' ) ) {
-	define( 'CMINB_USE_CACHE', true );
-}
 
 // Load the functionality.
 require_once CMINB_PATH . 'includes/Autoload.php';