+++ includes/Load.php
@@ -5,7 +5,8 @@
  * @package    Classic Menu in Navigation Block
  * @copyright  WebMan Design, Oliver Juhas
  *
- * @since  1.0.0
+ * @since    1.0.0
+ * @version  1.0.3
  */
 
 namespace WebManDesign\CMiNB;
@@ -18,7 +19,8 @@ class Load {
 	/**
 	 * Initialization.
 	 *
-	 * @since  1.0.0
+	 * @since    1.0.0
+	 * @version  1.0.3
 	 *
 	 * @return  void
 	 */
@@ -28,6 +30,7 @@ class Load {
 
 			Block::init();
 			Menu::init();
+			Options::init();
 
 			// Actions